Move the reference update thunks into the reference module
