+++ frontend/src/metabase/reference/databases/DatabaseDetail.tsx
@@ -6,11 +6,11 @@ import { t } from "ttag";
 import { LoadingAndErrorWrapper } from "metabase/common/components/LoadingAndErrorWrapper";
 import CS from "metabase/css/core/index.css";
 import { connect } from "metabase/redux";
-import { updateDatabase } from "metabase/redux/metadata";
 import Detail from "metabase/reference/components/Detail";
 import { EditHeader } from "metabase/reference/components/EditHeader";
 import EditableReferenceHeader from "metabase/reference/components/EditableReferenceHeader";
 import * as actions from "metabase/reference/reference";
+import { updateDatabase } from "metabase/reference/update-actions";
 import { getShallowFields as getFields } from "metabase/selectors/metadata";
 import type { User } from "metabase-types/api";
